+++ app/src/main/java/com/cz/sample/test/SampleMessageFragment.kt
@@ -26,11 +26,7 @@ class SampleMessageFragment : Fragment() {
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
         outButton.setOnClickListener {
-            println("System.out from fragment!")
-        }
-
-        outErrButton.setOnClickListener {
-            System.err.println("System.err from fragment!")
+            println("System.out from fragment!     .")
         }
     }
 }