+++ library/src/main/java/com/cz/android/sample/library/component/message/SampleMessageComponent.java
@@ -1,12 +1,16 @@
 package com.cz.android.sample.library.component.message;
 
+import android.text.Editable;
+import android.text.TextWatcher;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.FrameLayout;
+import android.widget.ScrollView;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
+import androidx.core.widget.NestedScrollView;
 import androidx.fragment.app.FragmentActivity;
 
 import com.cz.android.sample.component.ComponentContainer;
@@ -23,10 +27,11 @@ import java.util.Observer;
  */
 public class SampleMessageComponent implements ComponentContainer {
     private final WorkThread<String> workThread=new WorkThread("sample_work_thread");
-    private final SampleSystemConsole sampleSystemConsole=new SampleSystemConsole(workThread);
+    private final SampleSystemConsole sampleSystemConsole=new SampleSystemConsole();
 
     public SampleMessageComponent() {
         workThread.startService();
+        sampleSystemConsole.setup(workThread);
     }
 
     @Override
@@ -47,6 +52,7 @@ public class SampleMessageComponent implements ComponentContainer {
     @Override
     public void onCreatedView(FragmentActivity context, Object object, View view) {
         //If activity/fragment want to output message
+        final NestedScrollView scrollView=view.findViewById(R.id.sampleScrollView);
         final TextView messageView = view.findViewById(R.id.sampleMessageText);
         final Observer observer=new Observer() {
             @Override
@@ -59,6 +65,42 @@ public class SampleMessageComponent implements ComponentContainer {
                 });
             }
         };
+        final View scrollDownButton=view.findViewById(R.id.scrollDownButton);
+        scrollDownButton.setSelected(true);
+        final TextWatcher textWatcher = new TextWatcher() {
+            @Override
+            public void beforeTextChanged(CharSequence charSequence, int i, int i1, int i2) {
+            }
+            @Override
+            public void afterTextChanged(Editable editable) {
+            }
+            @Override
+            public void onTextChanged(CharSequence charSequence, int i, int i1, int i2) {
+                if(scrollDownButton.isSelected()){
+                    scrollView.fullScroll(ScrollView.FOCUS_DOWN);
+                }
+            }
+        };
+        messageView.addTextChangedListener(textWatcher);
+        scrollDownButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                scrollDownButton.setSelected(!scrollDownButton.isSelected());
+                if(scrollDownButton.isSelected()){
+                    messageView.addTextChangedListener(textWatcher);
+                } else {
+                    messageView.removeTextChangedListener(textWatcher);
+                }
+            }
+        });
+        View clearMessageButton=view.findViewById(R.id.clearMessageButton);
+        clearMessageButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                messageView.setText(null);
+            }
+        });
+
         workThread.addObserver(observer);
         //here we try to bind SampleMessageBindFragment
         SampleMessageBindFragment.injectIfNeededIn(context,workThread);