+++ document/changeLog.md
@@ -10,3 +10,20 @@ This is the first version of this project.
 
 all done
 
+
+### 1.0.1
+*时间:* 2020/2/2<br>
+*作者:* Jack chen
+Fixed a few bugs
+
+1. MessageComponent remove System.err pip line stream
+2. ViewPager default page limit in CompanionComponent
+
+### 1.0.2
+*时间:* 2020/2/6<br>
+*作者:* Jack chen
+Fixed a few bugs
+
+1. MessageComponent add two action buttons.
+    The first one is auto scroll to bottom.
+    The second one is clean all the message